requirements.txt
rembg==2.0.65
pillow>=10.0.1,<12.0.0
python-multipart==0.0.6
fastapi==0.103.1
onnxruntime>=1.17.0 